Modified _auth_tkt in line with paste upstream
Refreshed _auth_tkt to bring it in line with upstream paste auth_tkt.py
This brings with it support for non-MD5 digests.

All changes to _auth_tkt.py came from upstream.
It is not however a synchronised file, repoze variations have been
retained.

Changes were made to test__auth_tkt.py to allow tests to pass.
The number of arguments for calculate_digest() has changed requiring the
tests to be updated. This is an internal function and should not impact
users.